+++ lib/create_schedules/create_schedule.dart
@@ -4,16 +4,8 @@ import 'package:attendifyyy/authentication/user_preferences/user_preferences.dar
 import 'package:attendifyyy/bottom_nav_bar.dart';
 import 'package:flutter/material.dart';
 import 'package:http/http.dart' as http;
-
-List<String> dayOfWeekList = [
-  "Monday",
-  "Tuesday",
-  "Wednesday",
-  "Thursday",
-  "Friday",
-  "Saturday",
-  "Sunday"
-];
+import 'package:attendifyyy/create_schedules/create_schedule_form.dart';
+import 'package:attendifyyy/create_schedules/edit_schedule_form.dart';
 
 class ListOfSchedules extends StatefulWidget {
   const ListOfSchedules({super.key});
@@ -196,12 +188,20 @@ class ClassScheduleCard extends StatelessWidget {
                               const Icon(Icons.more_vert, color: Colors.white),
                           itemBuilder: (context) {
                             return [
-                              const PopupMenuItem<int>(
+                              PopupMenuItem<int>(
                                 value: 0,
                                 child: Row(children: [
-                                  Icon(Icons.edit, size: 18.0),
-                                  SizedBox(width: 5.0),
-                                  Text("Edit")
+                                  const Icon(Icons.edit, size: 18.0),
+                                  const SizedBox(width: 5.0),
+                                  TextButton(
+                                      onPressed: () {
+                                        showDialog(
+                                            context: context,
+                                            builder: (BuildContext context) =>
+                                                Dialog(
+                                                    child: EditSchedule()));
+                                      },
+                                      child: const Text("Edit")),
                                 ]),
                               ),
                               const PopupMenuItem<int>(
@@ -279,309 +279,3 @@ class ClassScheduleCard extends StatelessWidget {
   }
 }
 
-/*
-*
-* Using for creating the schedule dialog content
-*
-* */
-class CreateSchedule extends StatefulWidget {
-  const CreateSchedule({super.key});
-
-  @override
-  State<CreateSchedule> createState() => _CreateScheduleState();
-}
-
-class _CreateScheduleState extends State<CreateSchedule> {
-  final _subjectFormKey = GlobalKey<FormState>();
-  String? subjectNameValue;
-  String? sectionNameValue;
-  TimeOfDay startTime = TimeOfDay.now();
-  TimeOfDay endTime = TimeOfDay.now();
-  String? dayWeekValue;
-
-  List<dynamic> converted = [];
-
-  @override
-  void initState() {
-    super.initState();
-    getListOfSubjects();
-  }
-
-  Future<void> getListOfSubjects() async {
-    Map<String, dynamic>? teacherInfo =
-        await RememberUserPreferences.readUserInfo();
-
-    String? teacherId = teacherInfo?['teacher_id'];
-    if (teacherId != null && teacherId.isNotEmpty) {
-      final response = await http
-          .get(Uri.parse('${Api.listOfSubjects}?teacher_id=$teacherId'));
-      if (response.statusCode == 200) {
-        if (response.body.isNotEmpty) {
-          setState(() {
-            converted = jsonDecode(response.body);
-          });
-          // print("KANI SIYAAAA: ${converted[index]['subject_name']}");
-        } else {
-          ScaffoldMessenger.of(context)
-              .showSnackBar(SnackBar(content: Text("No subjects")));
-        }
-      } else {
-        ScaffoldMessenger.of(context)
-            .showSnackBar(SnackBar(content: Text("Failed to fetch subjects")));
-      }
-    } else {
-      print("Error: Teacher ID is null or empty");
-    }
-  }
-
-  //post newly created schedule to the database
-  Future<void> createSchedule() async {
-    Map<String, dynamic>? teacherInfo =
-        await RememberUserPreferences.readUserInfo();
-
-    String teacherId = teacherInfo?['teacher_id'];
-
-    print("Sulod sa start time: ${startTime}");
-    print('subject name sa FUNCTION $subjectNameValue');
-    print('section name sa FUNCTION $sectionNameValue');
-
-    final response = await http.post(Uri.parse(Api.createSchedule), body: {
-      'teacher_id': teacherId,
-      'subject_name': subjectNameValue,
-      'section_name': sectionNameValue,
-      'start_time':
-          "$startTime.", //I wrap it with double quote to convert it into string
-      'end_time': "$endTime",
-      'days_of_week': dayWeekValue
-    });
-    ScaffoldMessenger.of(context)
-        .showSnackBar(SnackBar(content: Text('${response.body}')));
-
-    print(response.body);
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      height: 480,
-      padding: const EdgeInsets.all(24.0),
-      child: ListView(
-        children: [
-          const Text('CREATE SCHEDULE',
-              style: TextStyle(
-                fontWeight: FontWeight.bold,
-                fontSize: 16.0,
-              )),
-          Form(
-              key: _subjectFormKey,
-              child: Column(children: [
-                const SizedBox(height: 20),
-                /*
-                *
-                * Dropdown for subject name
-                *
-                * */
-                DropdownButtonFormField<String>(
-                  validator: (value) {
-                    if (value == null || value.isEmpty) {
-                      return "This field is required.";
-                    }
-                    return null;
-                  },
-                  decoration: InputDecoration(
-                    labelText: "Subject",
-                    contentPadding: const EdgeInsets.all(16.0),
-                    //border style when its focus
-                    focusedBorder: const OutlineInputBorder(
-                        borderRadius: BorderRadius.all(Radius.circular(10.0)),
-                        borderSide:
-                            BorderSide(width: 2, color: Color(0xFF081631))),
-                    //border radius
-                    border: OutlineInputBorder(
-                      borderRadius: BorderRadius.circular(10),
-                    ),
-                  ),
-                  isExpanded: true, //set width to 100%
-                  // value: dayWeekValue,
-                  icon: const Icon(Icons.arrow_drop_down),
-                  items:
-                      converted.map<DropdownMenuItem<String>>((dynamic value) {
-                    String combinedValue =
-                        "${value['section_name']} - ${value['subject_name']}";
-                    return DropdownMenuItem(
-                        value: combinedValue, child: Text(combinedValue));
-                  }).toList(),
-                  onChanged: (String? value) {
-                    List<String> values = value!.split(RegExp(r'\s-\s'));
-                    print('SA SETSTATE ${values[0]}');
-                    setState(() {
-                      sectionNameValue = values[0];
-                      subjectNameValue = values[1];
-                      print('section name sa SETSTATE $sectionNameValue');
-                      print('subject name sa SETSTATE $subjectNameValue');
-                    });
-                  },
-                ),
-                const SizedBox(height: 20),
-                /*
-                *
-                * Time picker for start time
-                *
-                * */
-                OutlinedButton(
-                    style: ButtonStyle(
-                        padding: MaterialStateProperty.all(
-                            const EdgeInsets.symmetric(
-                                vertical: 20.0, horizontal: 14.0)),
-                        shape: MaterialStateProperty.all(
-                          RoundedRectangleBorder(
-                            borderRadius: BorderRadius.circular(10.0),
-                          ),
-                        )),
-                    onPressed: () async {
-                      final TimeOfDay? timeOfDay = await showTimePicker(
-                        context: context,
-                        initialTime: startTime,
-                        initialEntryMode: TimePickerEntryMode.dial,
-                      );
-                      if (timeOfDay != null) {
-                        setState(() {
-                          startTime = timeOfDay;
-                        });
-                      }
-                    },
-                    child: Row(
-                      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                      children: [
-                        Row(children: [
-                          const Text("Start Time:  ",
-                              style: TextStyle(color: Color(0xFF081631))),
-                          Text("${startTime.hour}:${startTime.minute}",
-                              style: const TextStyle(
-                                  color: Color(0xFF081631),
-                                  fontSize: 15.0,
-                                  fontWeight: FontWeight.bold)),
-                        ]),
-                        const Icon(Icons.schedule, color: Color(0xFF081631))
-                      ],
-                    )),
-                const SizedBox(height: 20),
-                /*
-                *
-                * Time picker for end time
-                *
-                * */
-                OutlinedButton(
-                    style: ButtonStyle(
-                        padding: MaterialStateProperty.all(
-                            const EdgeInsets.symmetric(
-                                vertical: 20.0, horizontal: 14.0)),
-                        shape: MaterialStateProperty.all(
-                          RoundedRectangleBorder(
-                            borderRadius: BorderRadius.circular(10.0),
-                          ),
-                        )),
-                    onPressed: () async {
-                      final TimeOfDay? timeOfDay = await showTimePicker(
-                        context: context,
-                        initialTime: endTime,
-                        initialEntryMode: TimePickerEntryMode.dial,
-                      );
-                      if (timeOfDay != null) {
-                        setState(() {
-                          endTime = timeOfDay;
-                        });
-                      }
-                    },
-                    child: Row(
-                      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                      children: [
-                        Row(children: [
-                          const Text("End Time:  ",
-                              style: TextStyle(color: Color(0xFF081631))),
-                          Text("${endTime.hour}:${endTime.minute}",
-                              style: const TextStyle(
-                                  color: Color(0xFF081631),
-                                  fontSize: 15.0,
-                                  fontWeight: FontWeight.bold)),
-                        ]),
-                        const Icon(Icons.schedule, color: Color(0xFF081631))
-                      ],
-                    )),
-                const SizedBox(height: 20),
-                /*
-                *
-                * Dropdown field for day of the week
-                *
-                * */
-                DropdownButtonFormField<String>(
-                  decoration: InputDecoration(
-                    labelText: "Days in a week",
-                    contentPadding: const EdgeInsets.all(16.0),
-                    //border style when its focus
-                    focusedBorder: const OutlineInputBorder(
-                        borderRadius: BorderRadius.all(Radius.circular(10.0)),
-                        borderSide:
-                            BorderSide(width: 2, color: Color(0xFF081631))),
-                    //border radius
-                    border: OutlineInputBorder(
-                      borderRadius: BorderRadius.circular(10),
-                    ),
-                  ),
-                  isExpanded: true, //set width to 100%
-                  // value: dayWeekValue,
-                  icon: const Icon(Icons.arrow_drop_down),
-                  items: dayOfWeekList
-                      .map<DropdownMenuItem<String>>((String value) {
-                    return DropdownMenuItem(
-                        value: value, child: Text("$value"));
-                  }).toList(),
-                  onChanged: (String? value) {
-                    // This is called when the user selects an item.
-                    setState(() {
-                      dayWeekValue = value!;
-                    });
-                  },
-                ),
-                const SizedBox(height: 20),
-                /*
-                *
-                * submit button
-                *
-                * */
-                TextButton(
-                    onPressed: () {
-                      print("pressed");
-                      //validate input fields
-                      if (_subjectFormKey.currentState!.validate()) {
-                        print("submitted");
-                        createSchedule();
-                      }
-                    },
-                    style: ButtonStyle(
-                      minimumSize: MaterialStateProperty.all<Size>(
-                          const Size.fromHeight(
-                              55)), //having height will make width 100%
-                      padding: MaterialStateProperty.all<EdgeInsets>(
-                          const EdgeInsets.symmetric(
-                              vertical: 14.0, horizontal: 44.0)),
-                      backgroundColor: MaterialStateProperty.all<Color>(
-                        const Color(0xFF081631),
-                      ),
-                      shape: MaterialStateProperty.all<RoundedRectangleBorder>(
-                          RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(10.0),
-                      )),
-                    ),
-                    child: const Text("Create",
-                        style: TextStyle(
-                            backgroundColor: Color(0xFF081631),
-                            fontSize: 16,
-                            fontWeight: FontWeight.bold,
-                            color: Colors.white)))
-              ])),
-        ],
-      ),
-    );
-  }
-}
